requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=dev.txt dev.in
#
appdirs==1.4.4            # via black, ptpython
asgiref==3.2.10           # via -r main.txt, django
attrs==19.3.0             # via black, pytest
bcrypt==3.1.7             # via paramiko
black==19.10b0            # via -r dev.in
certifi==2020.6.20        # via requests
cffi==1.14.0              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click==7.1.2              # via -r dev.in, black, pip-tools
cryptography==2.9.2       # via paramiko
django-environ==0.4.5     # via -r main.txt
django-extensions==3.0.2  # via -r dev.in
django-widget-tweaks==1.4.8  # via -r main.txt
django==3.0.7             # via -r main.txt
fabric==2.5.0             # via -r dev.in, patchwork
factory-boy==2.12.0       # via -r dev.in, pytest-factoryboy
faker==4.1.1              # via factory-boy
gunicorn==20.0.4          # via -r main.txt
idna==2.10                # via requests
inflection==0.5.0         # via pytest-factoryboy
invoke==1.4.1             # via fabric
jedi==0.17.1              # via ptpython
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
paramiko==2.7.1           # via fabric
parso==0.7.0              # via jedi
patchwork==1.0.1          # via -r dev.in
pathspec==0.8.0           # via black
pip-tools==5.2.1          # via -r dev.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.5     # via ptpython
ptpython==3.0.2           # via -r dev.in
py==1.8.2                 # via pytest
pycparser==2.20           # via cffi
pygments==2.6.1           # via ptpython
pynacl==1.4.0             # via paramiko
pyparsing==2.4.7          # via packaging
pytest-base-url==1.4.2    # via pytest-selenium
pytest-django==3.9.0      # via -r dev.in
pytest-factoryboy==2.0.3  # via -r dev.in
pytest-html==2.1.1        # via pytest-selenium
pytest-metadata==1.10.0   # via pytest-html
pytest-selenium==1.17.0   # via -r dev.in
pytest-variables==1.9.0   # via pytest-selenium
pytest==5.4.3             # via -r dev.in, pytest-base-url, pytest-django, pytest-factoryboy, pytest-html, pytest-metadata, pytest-selenium, pytest-variables
python-dateutil==2.8.1    # via faker
python-dotenv==0.14.0     # via -r dev.in
pytz==2020.1              # via -r main.txt, django
regex==2020.6.8           # via black
requests==2.24.0          # via pytest-base-url, pytest-selenium
selenium==3.141.0         # via -r dev.in, pytest-selenium
six==1.15.0               # via bcrypt, cryptography, django-extensions, packaging, pip-tools, pynacl, python-dateutil
sqlparse==0.3.1           # via -r main.txt, django
text-unidecode==1.3       # via faker
toml==0.10.1              # via black
typed-ast==1.4.1          # via black
urllib3==1.25.9           # via requests, selenium
wcwidth==0.2.5            # via prompt-toolkit, pytest

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
